# Metrics: Avoid defensive copies of MetricPoint struct