## Description

We have this pattern in our metric exporters where we do...

```csharp
foreach (ref readonly var metricPoint in metric.GetMetricPoints())
{
   long value = metricPoint.GetSumLong();
}
```

`MetricPoint` is a mutable struct:

https://github.com/open-telemetry/opentelemetry-dotnet/blob/a1e716716cd2c586119ea9a062e6cf88faccb63f/src/OpenTelemetry/Metrics/MetricPoint.cs#L27

When the compiler sees the `GetSumLong` invocation (for example) it doesn't know if that call is going to mutate the state of the local `metricPoint` so it will create a defensive copy. This can be avoided by declaring members which do not modify the state of the struct as `readonly`.

Here's a doc which goes over this in more detail: https://docs.microsoft.com/en-us/dotnet/csharp/write-safe-efficient-code#declare-readonly-members-for-mutable-structs

## Benchmarks

(I used `PrometheusSerializerBenchmarks` but this will be universal wherever `MetricPoint` is used as a `ref readonly`.)

### Before

|      Method | NumberOfSerializeCalls |          Mean |       Error |      StdDev | Allocated |
|------------ |----------------------- |--------------:|------------:|------------:|----------:|
| WriteMetric |                      1 |      5.324 us |   0.1063 us |   0.2781 us |         - |
| WriteMetric |                   1000 |  5,499.300 us | 109.2757 us | 112.2181 us |       6 B |
| WriteMetric |                  10000 | 54,968.431 us | 799.3971 us | 708.6447 us |     172 B |

### After

|      Method | NumberOfSerializeCalls |          Mean |       Error |      StdDev | Allocated |
|------------ |----------------------- |--------------:|------------:|------------:|----------:|
| WriteMetric |                      1 |      **4.427 us** |   0.0462 us |   0.0432 us |         - |
| WriteMetric |                   1000 |  **4,814.730 us** |  67.4480 us |  90.0411 us |      11 B |
| WriteMetric |                  10000 | **48,092.389 us** | 780.6455 us | 692.0219 us |      70 B |

## TODOs

* [X] Appropriate `CHANGELOG.md` updated for non-trivial changes