@utpilla 2022-03-17
Why are the allocation numbers changing randomly? Is this change supposed to reduce allocation as well? If yes, why does it go up from 6B to 11B for WriteMetric with 1000 calls?