@CodeBlanch 2022-03-17
@utpilla That is a good question I do not have an answer for. Should be no change at all to the allocations for this work. Even more then that, there should no allocations at all ever shown with this benchmark! Prometheus reuses a buffer so that it doesn't need allocations. Seems to be noise coming from _somewhere_. Something I would like to get to the bottom of, but prefer to not block this PR for that.